Pistol Charged Fire
Pistol can now fire a charged bolt even though nobody asked for that.

this is mostly just a test to make sure that held fire works; means we can implement it in the future for automatic weapons

additionally cleaned up some other parts of code; in the future bullets can probably extend the base classes PhysicalBullet and KinematicBullet.